:recycle: Refactor getNewPos and move to utils
