+++ app/containers/Lists/actions.js
@@ -4,6 +4,8 @@
  *
  */
 
+import getNewPos from 'utils/lib/getNewPos';
+
 import { ADD_LIST, CHANGE_POSITION } from './constants';
 
 export const addList = (title, idBoard) => ({
@@ -12,28 +14,8 @@ export const addList = (title, idBoard) => ({
   idBoard,
 });
 
-export const changePosList = (collection, source, destination) => {
-  const idList = collection[source].id;
-
-  const destinationPos = collection[destination].pos;
-  let itemAfterOrBeforeDestination = null;
-  let beforeOrAfterPos = destination > source ? 999999 : 0;
-
-  if (destination > source) {
-    itemAfterOrBeforeDestination = collection[destination + 1];
-    if (itemAfterOrBeforeDestination)
-      beforeOrAfterPos = itemAfterOrBeforeDestination.pos;
-  } else {
-    itemAfterOrBeforeDestination = collection[destination - 1];
-    if (itemAfterOrBeforeDestination)
-      beforeOrAfterPos = itemAfterOrBeforeDestination.pos;
-  }
-
-  const newPos = (destinationPos + beforeOrAfterPos) / 2;
-
-  return {
-    type: CHANGE_POSITION,
-    idList,
-    newPos,
-  };
-};
+export const changePosList = (collection, source, destination) => ({
+  type: CHANGE_POSITION,
+  idList: collection[source].id,
+  newPos: getNewPos(collection, source, destination),
+});
